feat: / 루트 페이지 자바스크립트 및 hbs 추가
