+++ src/main/java/chess/model/dto/RoomDto.java
@@ -1,0 +1,14 @@
+package chess.model.dto;
+
+public class RoomDto {
+    private String roomName;
+    private String password;
+
+    public String getRoomName() {
+        return roomName;
+    }
+
+    public String getPassword() {
+        return password;
+    }
+}
