+++ src/main/resources/templates/game.hbs
@@ -1,0 +1,103 @@
+<!DOCTYPE html>
+<html lang="ko">
+<head>
+    <meta charset="UTF-8">
+    <title>체스 게임</title>
+    <link href="/src/index.css" rel="stylesheet">
+    <script src="/src/game.js"></script>
+</head>
+<body>
+<h1> 체스 </h1>
+<div id="page">
+    <section>
+        <table id="chess-board">
+            <tr>
+                <td class="white-box" id="a8"></td>
+                <td class="black-box" id="b8"></td>
+                <td class="white-box" id="c8"></td>
+                <td class="black-box" id="d8"></td>
+                <td class="white-box" id="e8"></td>
+                <td class="black-box" id="f8"></td>
+                <td class="white-box" id="g8"></td>
+                <td class="black-box" id="h8"></td>
+            </tr>
+            <tr>
+                <td class="black-box" id="a7"></td>
+                <td class="white-box" id="b7"></td>
+                <td class="black-box" id="c7"></td>
+                <td class="white-box" id="d7"></td>
+                <td class="black-box" id="e7"></td>
+                <td class="white-box" id="f7"></td>
+                <td class="black-box" id="g7"></td>
+                <td class="white-box" id="h7"></td>
+            </tr>
+            <tr>
+                <td class="white-box" id="a6"></td>
+                <td class="black-box" id="b6"></td>
+                <td class="white-box" id="c6"></td>
+                <td class="black-box" id="d6"></td>
+                <td class="white-box" id="e6"></td>
+                <td class="black-box" id="f6"></td>
+                <td class="white-box" id="g6"></td>
+                <td class="black-box" id="h6"></td>
+            </tr>
+            <tr>
+                <td class="black-box" id="a5"></td>
+                <td class="white-box" id="b5"></td>
+                <td class="black-box" id="c5"></td>
+                <td class="white-box" id="d5"></td>
+                <td class="black-box" id="e5"></td>
+                <td class="white-box" id="f5"></td>
+                <td class="black-box" id="g5"></td>
+                <td class="white-box" id="h5"></td>
+            </tr>
+            <tr>
+                <td class="white-box" id="a4"></td>
+                <td class="black-box" id="b4"></td>
+                <td class="white-box" id="c4"></td>
+                <td class="black-box" id="d4"></td>
+                <td class="white-box" id="e4"></td>
+                <td class="black-box" id="f4"></td>
+                <td class="white-box" id="g4"></td>
+                <td class="black-box" id="h4"></td>
+            </tr>
+            <tr>
+                <td class="black-box" id="a3"></td>
+                <td class="white-box" id="b3"></td>
+                <td class="black-box" id="c3"></td>
+                <td class="white-box" id="d3"></td>
+                <td class="black-box" id="e3"></td>
+                <td class="white-box" id="f3"></td>
+                <td class="black-box" id="g3"></td>
+                <td class="white-box" id="h3"></td>
+            </tr>
+            <tr>
+                <td class="white-box" id="a2"></td>
+                <td class="black-box" id="b2"></td>
+                <td class="white-box" id="c2"></td>
+                <td class="black-box" id="d2"></td>
+                <td class="white-box" id="e2"></td>
+                <td class="black-box" id="f2"></td>
+                <td class="white-box" id="g2"></td>
+                <td class="black-box" id="h2"></td>
+            </tr>
+            <tr>
+                <td class="black-box" id="a1"></td>
+                <td class="white-box" id="b1"></td>
+                <td class="black-box" id="c1"></td>
+                <td class="white-box" id="d1"></td>
+                <td class="black-box" id="e1"></td>
+                <td class="white-box" id="f1"></td>
+                <td class="black-box" id="g1"></td>
+                <td class="white-box" id="h1"></td>
+            </tr>
+        </table>
+    </section>
+    <section id="game-info-section">
+        <button class="button" id="game-button" onclick="clickButton()"> start!</button>
+        <div id="turn-box"> 아직 게임 시작을 하지 않았습니다.</div>
+        <button id="quit" onclick="quit()"> 방 나가기</button>
+    </section>
+</div>
+</body>
+</html>
